Move multi-tokenizer event loop to better place
